完成个人信息非空判断

+++ work_system/src/main/java/com/idiotic/dao/CompanyDao.java
@@ -4,5 +4,5 @@ import com.idiotic.domain.system.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface CompanyDao extends JpaRepository<Company,Integer>, JpaSpecificationExecutor<Company> {
+public interface CompanyDao extends JpaRepository<Company,Long>, JpaSpecificationExecutor<Company> {
 }
