+++ work_system/src/main/java/com/idiotic/dao/MyInfoDao.java
@@ -4,6 +4,6 @@ import com.idiotic.domain.system.MyInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface MyInfoDao extends JpaRepository<MyInfo,Integer>, JpaSpecificationExecutor<MyInfo> {
-    MyInfo findByUserId(Integer id);
+public interface MyInfoDao extends JpaRepository<MyInfo,Long>, JpaSpecificationExecutor<MyInfo> {
+    MyInfo findByUserId(Long id);
 }
